Event Organizer => (Event Organizer)Layout/Index => Notification =>

+++ WebApplication1/Controllers/Event_OrganizersController.cs
@@ -7,6 +7,8 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using WebApplication1.Models;
+using WebApplication1.Utils;
+using static Dropbox.Api.TeamLog.SharedLinkAccessLevel;
 
 namespace WebApplication1.Controllers
 {
@@ -18,9 +20,14 @@ namespace WebApplication1.Controllers
         public ActionResult Index()
         {
             var event_Organizers = db.Event_Organizers.Include(e => e.Admin);
-            return View(event_Organizers.ToList());
+            return View(event_Organizers.Where(x => x.EventOrganizer_ID == BaseHelper.event_organizers.EventOrganizer_ID).ToList());
         }
 
+        public ActionResult IndexOrganizer()
+        {
+            var event_Organizer = db.Event_Organizers.Include(w => w.Admin);
+            return View(event_Organizer.ToList());
+        }
         // GET: Event_Organizers/Details/5
         public ActionResult Details(int? id)
         {
@@ -35,6 +42,16 @@ namespace WebApplication1.Controllers
             }
             return View(event_Organizers);
         }
+        public ActionResult activateaccount(int id)
+        {
+            Event_Organizers event_Organizers = db.Event_Organizers.Where(x => x.EventOrganizer_ID == id).FirstOrDefault();
+            event_Organizers.Status = 1;
+            MailProvider.SentfromMail(event_Organizers.EventOrganizer_Email, "Account Activation Mail!!", "Dear " + event_Organizers.EventOrganizer_Name + "!! Your account has been Activated by Admin. Check out Your account details.<br /> Regards EMS, Thanks!!");
+            TempData["success"] = event_Organizers.EventOrganizer_Email + " Account has been Activated";
+            db.Entry(event_Organizers).State = EntityState.Modified;
+            db.SaveChanges();
+            return RedirectToAction("IndexOrganizer", "Event_Organizers");
+        }
 
         // GET: Event_Organizers/Create
         public ActionResult Create()
@@ -48,7 +65,7 @@ namespace WebApplication1.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "EventOrganizer_ID,EventOrganizer_Name,EventOrganizer_Email,EventOrganizer_Password,EventOrganizer_Contact,EventOrganizer_Gender,EventOrganizer_Picture,Status,Admin_FID")] Event_Organizers event_Organizers)
+        public ActionResult Create(Event_Organizers event_Organizers)
         {
             if (ModelState.IsValid)
             {
@@ -61,6 +78,16 @@ namespace WebApplication1.Controllers
             return View(event_Organizers);
         }
 
+        public ActionResult blockaccount(int id)
+        {
+            Event_Organizers o = db.Event_Organizers.Where(x => x.EventOrganizer_ID == id).FirstOrDefault();
+            o.Status = -1;
+            MailProvider.SentfromMail(o.EventOrganizer_Email, "Account Blockage Mail!!", "Dear " + o.EventOrganizer_Name + "!! Your account has been Blocked by Admin. Check out Your account details.<br /> Regards EMS, Thanks!!");
+            TempData["success"] = o.EventOrganizer_Email + " Account has been blocked";
+            db.Entry(o).State = EntityState.Modified;
+            db.SaveChanges();
+            return RedirectToAction("IndexOrganizer", "Event_Organizers");
+        }
         // GET: Event_Organizers/Edit/5
         public ActionResult Edit(int? id)
         {
@@ -82,7 +109,7 @@ namespace WebApplication1.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "EventOrganizer_ID,EventOrganizer_Name,EventOrganizer_Email,EventOrganizer_Password,EventOrganizer_Contact,EventOrganizer_Gender,EventOrganizer_Picture,Status,Admin_FID")] Event_Organizers event_Organizers)
+        public ActionResult Edit(Event_Organizers event_Organizers)
         {
             if (ModelState.IsValid)
             {
@@ -115,11 +142,15 @@ namespace WebApplication1.Controllers
         public ActionResult DeleteConfirmed(int id)
         {
             Event_Organizers event_Organizers = db.Event_Organizers.Find(id);
-            db.Event_Organizers.Remove(event_Organizers);
-            db.SaveChanges();
-            return RedirectToAction("Index");
-        }
+            if (ModelState.IsValid)
+            {
+                event_Organizers.Status = 0;
+                db.Event_Organizers.Remove(event_Organizers);
+                db.SaveChanges();
+            }
 
+            return RedirectToAction("Index", "Home");
+        }
         protected override void Dispose(bool disposing)
         {
             if (disposing)
@@ -127,6 +158,129 @@ namespace WebApplication1.Controllers
                 db.Dispose();
             }
             base.Dispose(disposing);
+        }
+
+        //Reset Password
+        [HttpPost]
+        public ActionResult Forgotpasword(string email)
+        {
+            var v = db.Event_Organizers.Where(x => x.EventOrganizer_Email == email).FirstOrDefault();
+            if (v == null)
+            {
+                TempData["msg"] = "invalid email";
+                return RedirectToAction("login");
+            }
+            Random random = new Random();
+            int code = random.Next(1001, 9999);
+            Session["code"] = code;
+            Session["userforgotpassword"] = v;
+            MailProvider.Sentforgotpassmail(v.EventOrganizer_Email, code);
+            return RedirectToAction("CodeVerify");
+
+
+        }
+        public ActionResult CodeVerify()
+        {
+            return View();
+        }
+        [HttpPost]
+        public ActionResult CodeVerify(int code)
+        {
+            int sentcode = (int)Session["code"];
+            if (code == sentcode)
+            {
+                TempData["msg"] = "Code validate!! Set your new password";
+                return RedirectToAction("NewPassword");
+            }
+            TempData["error"] = "invalid code. \n Try Again!";
+            return View();
+        }
+        public ActionResult NewPassword()
+        {
+            return View();
+        }
+        [HttpPost]
+        public ActionResult NewPassword(string password)
+        {
+            var user = (Event_Organizers)Session["userforgotpassword"];
+            var v = db.Event_Organizers.Where(x => x.EventOrganizer_Email == user.EventOrganizer_Email).FirstOrDefault();
+            v.EventOrganizer_Password = password;
+            db.Entry(v).State = EntityState.Modified;
+            db.SaveChanges();
+            TempData["msg"] = "Successfully changed password!";
+            return RedirectToAction("login");
+        }
+        // send password to email
+        public ActionResult Forgotpassword()
+        {
+            return View();
+        }
+        [HttpPost]
+        public ActionResult Forgotpassword(Event_Organizers event_Organizers)
+        {
+
+            Event_Organizers var = db.Event_Organizers.Where(x => x.EventOrganizer_Email == event_Organizers.EventOrganizer_Email && event_Organizers.EventOrganizer_Name == event_Organizers.EventOrganizer_Name).FirstOrDefault();
+            if (var != null)
+            {
+                MailProvider.SentfromMail(var.EventOrganizer_Email, "Fogotton Password!!", "Dear " + event_Organizers.EventOrganizer_Name + "!! , Your Password is  " + var.EventOrganizer_Password + "\n\n\nThanks & Regards\nEMS Team<br /> Thanks");
+                TempData["send"] = "Your Password has been sent to Your Registered Email Address.\n Please! Check Your Mail Inbox";
+            }
+            else
+            {
+                TempData["send"] = "Either Your Username is Not Valid or Email Not Registered";
+
+            }
+            return View();
+        }
+
+        public ActionResult login()
+        {
+            return View();
+        }
+        [HttpPost]
+        public ActionResult login(Event_Organizers event_Organizers)
+        {
+
+            int v = db.Event_Organizers.Where(x => x.EventOrganizer_Email == event_Organizers.EventOrganizer_Email && event_Organizers.EventOrganizer_Password == event_Organizers.EventOrganizer_Password).Count();
+            Event_Organizers var = db.Event_Organizers.Where(x => event_Organizers.EventOrganizer_Email == event_Organizers.EventOrganizer_Email && event_Organizers.EventOrganizer_Password == event_Organizers.EventOrganizer_Password).FirstOrDefault();
+            if (v > 0) {
+
+
+                if (var.Status == 1)
+                {
+                    BaseHelper.event_organizers = var;
+                    return RedirectToAction("IndexOrganizer", "home");
+
+                }
+
+                else if (var.Status == -1)
+                {
+                    ViewBag.message = "Your account has been blocked by admin";
+                    return View();
+                }
+                else if (var.Status == 0)
+                {
+                    ViewBag.message = "Your account has been deleted.Please send Mail to admin to Re-activate it.";
+                    return View();
+                }
+                else
+                {
+                    ViewBag.message = "Your account request has not been processed by admin";
+                    return View();
+                }
+            }
+            else
+            {
+                ViewBag.message = "Invalid Email Or Password";
+                return View();
+            }
+
+        }
+        public ActionResult logout()
+        {
+            BaseHelper.event_organizers = null;
+            return RedirectToAction("Index", "home");
+
         }
     }
 }
